Сверстано подробное окно мои наряды без реализации логики и кнопок, пофиксил баги в карточках при переносе текста, сверстал окно связанные согласования
