+++ .gitignore
@@ -9,6 +9,7 @@
 .history
 .svn/
 migrate_working_dir/
+*.txt
 
 # IntelliJ related
 *.iml
