+++ lib/widgets/main.dart
@@ -1,9 +1,12 @@
-import 'package:firstapp/authorization/authorization.dart';
-import 'package:firstapp/myOutfitsWindow/detailedOutfit.dart';
 import 'package:flutter/material.dart';
-import 'package:firstapp/myOutfitsWindow/myOutfitsWindow.dart';
-import 'package:firstapp/myApprovalsWindow/myApprovalsWindow.dart';
-import 'package:firstapp/myRequestsWindow/myRequestsWindow.dart';
+
+import 'authorization/authorization.dart';
+import 'myApprovalsWindow/detailedApprovals.dart';
+import 'myApprovalsWindow/myApprovalsWindow.dart';
+import 'myOutfitsWindow/detailedOutfit.dart';
+import 'myOutfitsWindow/myOutfitsWindow.dart';
+import 'myOutfitsWindow/relatedApprovals.dart';
+import 'myRequestsWindow/myRequestsWindow.dart';
 
 void main() {
   runApp( const Application());
@@ -12,7 +15,6 @@ void main() {
 
 class Application extends StatelessWidget {
   const Application({Key? key}) : super(key: key);
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -24,6 +26,8 @@ class Application extends StatelessWidget {
         '/myRequestsWindow': (context) => const MyRequestsWindow(),
         '/myApprovalsWindow': (context) => const MyApprovalsWindow(),
         '/myOutfitsWindow/detailedOutfit': (context) => const DetailedOutfit(),
+        '/myApprovalsWindow/detailedApprovals': (context) => const DetailedApprovals(),
+        '/myOutfitsWindow/detailedOutfit/relatedApprovals': (context) => const RelatedApprovals(),
       },
     );
   }
