+++ pubspec.yaml
@@ -32,6 +32,7 @@ dependencies:
     sdk: flutter
   sqflite:
   path:
+  pedantic_mono: any
 
 
   # The following adds the Cupertino Icons font to your application.
